+++ packages/base/src/components/Modal/providers/ModalOptionContext.ts
@@ -2,13 +2,19 @@ import { noop } from '@shoplflow/utils';
 import { createContext } from 'react';
 
 export type ModalOptionState = {
-  heightToDeduct: number;
-  setHeightToDeduct: (height: number) => void;
-  clearHeightToDeduct: () => void;
+  topHeight: number;
+  bottomHeight: number;
+  setTopHeight: (height: number) => void;
+  setBottomHeight: (height: number) => void;
+  clearTopHeight: () => void;
+  clearBottomHeight: () => void;
 };
 
 export const ModalOptionContext = createContext<ModalOptionState>({
-  heightToDeduct: 0,
-  setHeightToDeduct: noop,
-  clearHeightToDeduct: noop,
+  topHeight: 0,
+  bottomHeight: 0,
+  setTopHeight: noop,
+  setBottomHeight: noop,
+  clearTopHeight: noop,
+  clearBottomHeight: noop,
 });