+++ packages/base/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @shoplflow/base
 
+## 0.45.17
+
+### Patch Changes
+
+- Updated dependencies [[`d121c7ae2df2b79aaf069229603c35ed5e8af8fe`](https://github.com/shopl/shoplflow/commit/d121c7ae2df2b79aaf069229603c35ed5e8af8fe)]:
+  - @shoplflow/shopl-assets@0.12.36
+
+## 0.45.16
+
+### Patch Changes
+
+- [#716](https://github.com/shopl/shoplflow/pull/716) [`47371a78b23a3969e256918c4fecd0740c84f524`](https://github.com/shopl/shoplflow/commit/47371a78b23a3969e256918c4fecd0740c84f524) Thanks [@velo-kim](https://github.com/velo-kim)! - update(base): Modal Top, Bottom의 높이 변경 ResizeObserver 추가
+
+- Updated dependencies [[`47371a78b23a3969e256918c4fecd0740c84f524`](https://github.com/shopl/shoplflow/commit/47371a78b23a3969e256918c4fecd0740c84f524)]:
+  - @shoplflow/shopl-assets@0.12.35
+
+## 0.45.15
+
+### Patch Changes
+
+- [#715](https://github.com/shopl/shoplflow/pull/715) [`1e16cf193adff5ef44b5a64c9cf787a194f3f2f3`](https://github.com/shopl/shoplflow/commit/1e16cf193adff5ef44b5a64c9cf787a194f3f2f3) Thanks [@shopl-kevin](https://github.com/shopl-kevin)! - input leftsource 추가
+
+- [#713](https://github.com/shopl/shoplflow/pull/713) [`18c026569b3fb9c17968595abe85d0929c434f05`](https://github.com/shopl/shoplflow/commit/18c026569b3fb9c17968595abe85d0929c434f05) Thanks [@shopl-kevin](https://github.com/shopl-kevin)! - modal container hasChangeAnimation 시 height 애니메이션도 적용
+
+- Updated dependencies [[`774c1ebe97137327c6f934c7f5cda02b165e147f`](https://github.com/shopl/shoplflow/commit/774c1ebe97137327c6f934c7f5cda02b165e147f)]:
+  - @shoplflow/shopl-assets@0.12.34
+
 ## 0.45.14
 
 ### Patch Changes