+++ packages/shopl-assets/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @shoplflow/shopl-assets
 
+## 0.12.36
+
+### Patch Changes
+
+- [#719](https://github.com/shopl/shoplflow/pull/719) [`d121c7ae2df2b79aaf069229603c35ed5e8af8fe`](https://github.com/shopl/shoplflow/commit/d121c7ae2df2b79aaf069229603c35ed5e8af8fe) Thanks [@github-actions](https://github.com/apps/github-actions)! - icon 추가
+
+## 0.12.35
+
+### Patch Changes
+
+- [#716](https://github.com/shopl/shoplflow/pull/716) [`47371a78b23a3969e256918c4fecd0740c84f524`](https://github.com/shopl/shoplflow/commit/47371a78b23a3969e256918c4fecd0740c84f524) Thanks [@velo-kim](https://github.com/velo-kim)! - update(base): Modal Top, Bottom의 높이 변경 ResizeObserver 추가
+
+## 0.12.34
+
+### Patch Changes
+
+- [#712](https://github.com/shopl/shoplflow/pull/712) [`774c1ebe97137327c6f934c7f5cda02b165e147f`](https://github.com/shopl/shoplflow/commit/774c1ebe97137327c6f934c7f5cda02b165e147f) Thanks [@github-actions](https://github.com/apps/github-actions)! - icon 추가
+
 ## 0.12.33
 
 ### Patch Changes